Create a ModelInfo class and move some logic there
