+++ datadog_checks_dev/tests/tooling/configuration/consumers/model/test_deprecations.py
@@ -3,7 +3,7 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import pytest
 
-from ...utils import get_model_consumer, normalize_yaml
+from ...utils import get_model_consumer
 
 pytestmark = [pytest.mark.conf, pytest.mark.conf_consumer, pytest.mark.conf_consumer_model]
